[desk-tool] Fix styling of referring documents list
