--- packages/@sanity/desk-tool/src/components/styles/ReferringDocumentLink.css
@@ -1,3 +1,0 @@
-.root {
-  color: inherit;
-}
