+++ packages/@sanity/desk-tool/src/components/styles/ReferringDocumentsList.css
@@ -1,0 +1,11 @@
+
+.root {
+  composes: root from 'part:@sanity/components/lists/default-style';
+}
+.item {
+  composes: lineBetween from 'part:@sanity/components/lists/default-item-style';
+  padding: 1px 0;
+}
+.link {
+  color: inherit;
+}
